@@ -1,16 +1,16 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         applicationId "com.trayis.simplimvvm.app"
         minSdkVersion 17
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
@@ -40,13 +40,13 @@ task bintray() {
 }
 
 dependencies {
-    implementation "com.android.support:appcompat-v7:$androidSupportLibrary"
-    implementation "com.android.support:support-annotations:$androidSupportLibrary"
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.annotation:annotation:1.0.0'
 
-    implementation "com.android.support.constraint:constraint-layout:1.1.2"
+    implementation "androidx.constraintlayout:constraintlayout:2.0.0-alpha2"
 
-    implementation "android.arch.lifecycle:extensions:$androidArch"
-    implementation "android.arch.lifecycle:viewmodel:$androidArch"
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel:2.0.0'
 
     implementation project(":simpliMvvm")
     implementation project(":simpliMvvmAnnotation")
@@ -54,7 +54,7 @@ dependencies {
 
     testImplementation "junit:junit:4.12"
 
-    androidTestImplementation "com.android.support.test:runner:1.0.2"
-    androidTestImplementation "com.android.support.test.espresso:espresso-core:3.0.2"
+    androidTestImplementation "androidx.test:runner:1.1.0-beta01"
+    androidTestImplementation "androidx.test.espresso:espresso-core:3.1.0-beta02"
     testAnnotationProcessor project(":simpliMvvmAnnotation")
 }
