@@ -7,15 +7,15 @@ def siteUrl = 'https://gitlab.com/trayis/android/SimpiMVVM'      // Homepage URL
 def gitUrl = 'https://gitlab.com/trayis/android/SimpiMVVM.git'   // Git repository URL
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 17
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
     }
 
@@ -44,14 +44,14 @@ android {
 
 dependencies {
 
-    implementation "com.android.support:appcompat-v7:$androidSupportLibrary"
+    implementation 'androidx.appcompat:appcompat:1.0.0'
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
 
     testImplementation "junit:junit:4.12"
 
-    androidTestImplementation "com.android.support.test:runner:1.0.2"
-    androidTestImplementation "com.android.support.test.espresso:espresso-core:3.0.2"
+    androidTestImplementation "androidx.test:runner:1.0.0"
+    androidTestImplementation "androidx.test.espresso:espresso-core:3.1.0-beta02"
 
 }
 
