@@ -8,16 +8,16 @@ def siteUrl = 'https://gitlab.com/trayis/android/SimpiMVVM'      // Homepage URL
 def gitUrl = 'https://gitlab.com/trayis/android/SimpiMVVM.git'   // Git repository URL
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 17
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
@@ -45,16 +45,16 @@ android {
 
 dependencies {
 
-    implementation "com.android.support:appcompat-v7:$rootProject.androidSupportLibrary"
-    implementation "com.android.support:support-annotations:$rootProject.androidSupportLibrary"
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.annotation:annotation:1.0.0'
 
-    implementation "android.arch.lifecycle:extensions:$rootProject.androidArch"
-    implementation "android.arch.lifecycle:viewmodel:$rootProject.androidArch"
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel:2.0.0'
 
     testImplementation "junit:junit:4.12"
 
-    androidTestImplementation "com.android.support.test:runner:1.0.2"
-    androidTestImplementation "com.android.support.test.espresso:espresso-core:3.0.2"
+    androidTestImplementation "androidx.test:runner:1.1.0-beta01"
+    androidTestImplementation "androidx.test.espresso:espresso-core:3.1.0-beta02"
 }
 
 task prValidation {
@@ -65,7 +65,7 @@ task prValidation {
 }
 
 group = "com.trayis"    // Maven Group ID for the artifact
-version = "0.0.4"       // This is the library version used when deploying the artifact
+version = "1.0.0"       // This is the library version used when deploying the artifact
 
 install {
     repositories.mavenInstaller {
