@@ -1,10 +1,11 @@
 package com.trayis.simplikmvvm.ui
 
-import android.support.annotation.LayoutRes
+import androidx.annotation.LayoutRes
 
 interface Simpli {
 
     @get:LayoutRes
+
     val layoutResourceId: Int
 
     val modelVariable: Int
