@@ -1,9 +1,9 @@
 package com.trayis.simplikmvvm.ui
 
-import android.databinding.DataBindingUtil
-import android.databinding.ViewDataBinding
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
 import com.trayis.simplikmvvm.utils.Logging
 import com.trayis.simplikmvvm.utils.SimpliProviderUtil
 import com.trayis.simplikmvvm.viewmodel.SimpliViewModel
