@@ -1,11 +1,11 @@
 package com.trayis.simplimvvm.ui;
 
-import android.databinding.DataBindingUtil;
-import android.databinding.ViewDataBinding;
+import androidx.databinding.DataBindingUtil;
+import androidx.databinding.ViewDataBinding;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -16,11 +16,11 @@ import com.trayis.simplimvvm.viewmodel.SimpliViewModel;
 
 import java.util.InvalidPropertiesFormatException;
 
-public abstract class SimpliFragment<B extends ViewDataBinding, V extends SimpliViewModel> extends Fragment implements Simpli {
+public abstract class SimpliFragment<B extends ViewDataBinding> extends Fragment implements SimpliBase {
 
     protected final String TAG = getClass().getSimpleName();
 
-    private V mViewModel;
+    private SimpliViewModel[] mViewModel;
     private B mBinding;
 
     @Override
@@ -31,10 +31,10 @@ public abstract class SimpliFragment<B extends ViewDataBinding, V extends Simpli
     }
 
     @SuppressWarnings("unchecked")
-    private V getViewModel() {
+    private SimpliViewModel[] getViewModel() {
         if (mViewModel == null) {
             try {
-                mViewModel = (V) SimpliProviderUtil.getInstance().getProvider().getViewModel(this);
+                mViewModel = SimpliProviderUtil.getInstance().getProvider().getViewModels(this);
             } catch (InvalidPropertiesFormatException e) {
                 Logging.e(TAG, e.getMessage(), e);
             }
@@ -52,11 +52,11 @@ public abstract class SimpliFragment<B extends ViewDataBinding, V extends Simpli
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
         mBinding.setLifecycleOwner(this);
-        int modelVariable = getModelVariable();
-        if (modelVariable > 0) {
-            mBinding.setVariable(modelVariable, mViewModel);
+        if (mViewModel != null) {
+            for (SimpliViewModel viewModel : mViewModel) {
+                viewModel.onCreate();
+            }
         }
-        mViewModel.onCreate();
         mBinding.executePendingBindings();
     }
 
