@@ -1,23 +1,25 @@
 package com.trayis.simplimvvm.app.ui;
 
-import com.trayis.simplimvvm.app.BR;
 import com.trayis.simplimvvm.app.R;
 import com.trayis.simplimvvm.app.databinding.ActivityMainBinding;
 import com.trayis.simplimvvm.app.ui.model.MainViewModel;
 import com.trayis.simplimvvm.ui.SimpliActivity;
+import com.trayis.simplimvvmannotation.SimpliInject;
 import com.trayis.simplimvvmannotation.SimpliViewComponent;
 
 @SimpliViewComponent
-public class MainActivity extends SimpliActivity<ActivityMainBinding, MainViewModel> {
+public class MainActivity extends SimpliActivity<ActivityMainBinding> {
 
-    @Override
-    public int getModelVariable() {
-        return BR.viewModel;
-    }
+    @SimpliInject
+    MainViewModel viewModel;
 
     @Override
     public int getLayoutResourceId() {
         return R.layout.activity_main;
     }
 
+    @Override
+    public void setViewModelsToView() {
+        mBinding.setViewModel(viewModel);
+    }
 }
