@@ -1,6 +1,6 @@
 package com.trayis.simplimvvm.utils;
 
-import com.trayis.simplimvvm.ui.Simpli;
+import com.trayis.simplimvvm.ui.SimpliBase;
 import com.trayis.simplimvvm.viewmodel.SimpliViewModel;
 
 import java.util.InvalidPropertiesFormatException;
@@ -9,8 +9,8 @@ import java.util.InvalidPropertiesFormatException;
  * Created by mudesai on 10/9/17.
  */
 
-public interface SimpliMvvmProvider<V extends SimpliViewModel> {
+public interface SimpliMvvmProvider {
 
-    V getViewModel(Simpli simpli) throws InvalidPropertiesFormatException;
+    SimpliViewModel[] getViewModels(SimpliBase simpli) throws InvalidPropertiesFormatException;
 
 }
